create KeyMg

+++ WinStudy/Object.cpp
@@ -23,10 +23,10 @@ RECT Object::GetRect()
 	vec2 _nowPos = GetPos();
 	vec2 _nowScale = GetScale();
 
-	RECT rt = {_nowPos.x - _nowScale.x / 2,
-				_nowPos.y - _nowScale.y / 2 ,
-				_nowPos.x + _nowScale.x / 2 ,
-				_nowPos.y + _nowScale.y / 2 };
+	RECT rt = {(int)(_nowPos.x - _nowScale.x / 2),
+				(int)(_nowPos.y - _nowScale.y / 2) ,
+				(int)(_nowPos.x + _nowScale.x / 2) ,
+				(int)(_nowPos.y + _nowScale.y / 2) };
 
 	return rt;
 }
